Stable update finally
Hey, this can now run perfectly fine, but takes a stupid amount of time.
I decided to just have it this way, and I'm going to work to thread it
with the go threads in the future (mostly just for the requests) to try
and speed it up. The main problem being the market_history and not being
able to grab certain dates